refactor RouterTest
